Send from0to1023

+++ ServoConfigurator/RxTx_dataStructures.h
@@ -2,14 +2,14 @@ struct TX_DATA_STRUCTURE{
   //put your variable definitions here for the data you want to send
   //THIS MUST BE EXACTLY THE SAME ON THE OTHER ARDUINO
   
-    int16_t devType;  // mode:  0 = fourSticksController (8 chanels) ,   1 = ServoConfigurator (16 chanels) , 2 = MinMaxServoConfig (min max for 2 chanels)
+    uint16_t devType;  // mode:  0 = fourSticksController (8 chanels) ,   1 = ServoConfigurator (16 chanels) , 2 = MinMaxServoConfig (min max for 2 chanels)
 
     byte servoSet;
 
-    byte s1min;
-    byte s1mid;
-    byte s1curr;
-    byte s1max;
+    uint16_t s1min;
+    uint16_t s1mid;
+    uint16_t s1curr;
+    uint16_t s1max;
 
     byte fireBtn1;
